refactor(utils): move the legacy end-reason mapping to utils
It is a pure mapping with no socket-manager state, so it sits with the
other small helpers.
