+++ src/utils/stream-end.test.ts
